app/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Databricks
databricks-sdk>=0.27.0
databricks-sql-connector>=3.4.0
databricks-langchain>=0.1.0

# LangChain
langchain>=0.2.0
langchain-core>=0.2.0

# HTTP clients (live data tools)
requests>=2.31.0

# Garmin
garminconnect>=0.2.20

# Pydantic (FastAPI dep, pin for compatibility)
pydantic>=2.6.0
